PF-468 add an unsupported notebooks resource type. (#97)
We want to start working with AI Notebook instances, including publishing them to Janitor. For now, we want Janitor to be able collect the notebook cloud resource uids as errors without knowing how to clean them up. Without this change, the Janitor cannot store the cloud resources with its DAO as the resource type is unknown.
Later, we will add support for deleting the notebooks.

This should have been done when we bumped the janitor client version to include the new CloudResourceUid subfield for notebooks in #95