PF-194 Succeed when deleting a notebook in delete requested project. (#104)
Handle a special case where the project containing a notebook instance is deleted before the notebook instance is deleted. When this happens, the notebook instance will be deleted.

This does not solve the case when the project was deleted so long ago that it has been completed deleted, but I expect this to still resolve many of the recent and incoming notebook errors.

Refactor Janitor to provide CRL COW beans instead of creating them from ClientConfig each time they're needed.